backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.2
PyPDF2==3.0.1
requests==2.31.0
aiohttp==3.9.1
web3==6.11.3
python-dateutil==2.8.2
